updated market state type

+++ common/market_state.h
@@ -1,9 +1,11 @@
+#include <cstdint>
+
 namespace MarketState
 {
-constexpr int OPEN       = 1;
-constexpr int CLOSED     = 2;
-constexpr int PRE_OPEN   = 1;
-constexpr int HALT       = 1;
-constexpr int UNKNOWN    = 1;
-constexpr int POST_CLOSE = 1;
+constexpr uint16_t OPEN       = 1;
+constexpr uint16_t CLOSED     = 2;
+constexpr uint16_t PRE_OPEN   = 1;
+constexpr uint16_t HALT       = 1;
+constexpr uint16_t UNKNOWN    = 1;
+constexpr uint16_t POST_CLOSE = 1;
 }; // namespace MarketState
